Move all middlewares to store folder
