--- src/middlewares/index.js
@@ -1,2 +1,0 @@
-export apiMiddleware from './apiMiddleware';
-export firebaseMiddleware from './firebaseMiddleware';
